DDDDDDDDDDDD

+++ gerenciador_consultas/appointments/views.py
@@ -40,30 +40,47 @@ def my_list_view(request):
     if is_patient := hasattr(request.user, 'patient'):
         context['patient'] = request.user.patient
         context['is_patient'] = is_patient
-        context['appointments'] =context['patient'].appointments.all
+        context['appointments'] = context['patient'].appointments.all
     if is_professional :=  hasattr(request.user, 'professional'):
         context['professional'] = request.user.professional
         context['is_professional'] = is_professional
-        context['appointments'] =context['professional'].appointments.all
+        context['appointments'] = context['professional'].appointments.all
     
         return render(request, 'list.html', context)
 
+@login_required
 def edit_view(request, id):
-    appointment = get_object_or_404(Appointment,id=id)
-    if request.method == 'POST':
-        form = AppointmentForm(request.POST, instance=appointment)
-        if form.is_valid():
-            form.save()
-            return redirect('appointments:list')
-    else:
-        form = AppointmentForm(instance=appointment)
+    if request.user.is_staff:
+        appointment = get_object_or_404(Appointment,id=id)
+        if request.method == 'POST':
+            form = AppointmentForm(request.POST, instance=appointment)
+            if form.is_valid():
+                form.save()
+                return redirect('appointments:list')
+        else:
+            form = AppointmentForm(instance=appointment)
 
     return render(request,'edit.html',{'form':form, 'appointment':appointment,})
 
+@login_required
+def add_view(request):
+    if request.user.is_staff:
+        if request.method == 'POST':
+            form = AppointmentForm(data=request.POST)
+            if form.is_valid():
+                form.save()
+                return redirect('appointments:list')
+        else:
+            form = AppointmentForm()
+
+    return render(request,'edit.html',{'form':form})
+
+@login_required
 def delete_view(request, id):
-    appointment = get_object_or_404(Appointment,id=id)
-    if appointment:
-        appointment.delete()
+    if request.user.is_staff:
+        appointment = get_object_or_404(Appointment,id=id)
+        if appointment:
+            appointment.delete()
     return redirect('appointments:list')
 
 def index_view(request):
